separate bridge test into dataset & test, fix package.json requires
