+++ libs/coin-modules/coin-stacks/package.json
@@ -110,7 +110,10 @@
     "invariant": "^2.2.4",
     "ripple-address-codec": "^5.0.0",
     "ripple-binary-codec": "^1.3.0",
-    "rxjs": "^7.8.1"
+    "rxjs": "^7.8.1",
+    "@stacks/network": "6.10.0",
+    "@stacks/transactions": "6.11.0",
+    "@zondax/ledger-stacks": "^1.0.2"
   },
   "devDependencies": {
     "@faker-js/faker": "^8.4.1",
