+++ libs/ledger-live-common/package.json
@@ -188,8 +188,6 @@
     "@ledgerhq/wallet-api-core": "^1.13.0",
     "@ledgerhq/wallet-api-exchange-module": "workspace:^",
     "@ledgerhq/wallet-api-server": "^1.6.0",
-    "@stacks/network": "6.10.0",
-    "@stacks/transactions": "6.11.0",
     "@stricahq/typhonjs": "^2.0.0",
     "@taquito/ledger-signer": "^20.0.0",
     "@ton-community/ton-ledger": "^7.0.1",
@@ -204,7 +202,6 @@
     "@zondax/ledger-cosmos-js": "3.0.3",
     "@zondax/ledger-filecoin": "^0.11.2",
     "@zondax/ledger-icp": "^0.7.0",
-    "@zondax/ledger-stacks": "^1.0.2",
     "async": "^3.2.3",
     "axios": "1.7.7",
     "bech32": "^1.1.3",
