+++ libs/ledger-live-common/src/families/stacks/bridge.integration.test.ts
@@ -1,0 +1,5 @@
+import "../../__tests__/test-helpers/setup";
+import { testBridge } from "../../__tests__/test-helpers/bridge";
+import { dataset } from "@ledgerhq/coin-stacks/test";
+
+testBridge(dataset);
