+++ package.json
@@ -76,6 +76,7 @@
     "coin:ton": "pnpm --filter coin-ton",
     "coin:solana": "pnpm --filter coin-solana",
     "coin:hedera": "pnpm --filter coin-hedera",
+    "coin:stacks": "pnpm --filter coin-stacks",
     "coin:icon": "pnpm --filter coin-icon",
     "coin:stellar": "pnpm --filter coin-stellar",
     "coin:tezos": "pnpm --filter coin-tezos",
